feat(database)!: differentiate id and urlName of Section table
Add a `urlName` column to the `Section` table, and change the `id` column to an auto-incrementing Int instead. This makes the datatypes more consistent across the tables (and potentially gives us the option to edit a urlName after the section is created -- but that functionality shouldn't be needed often).

Also add a `database` scope to `.vscode/settings.json` for database-related changes like this one.

BREAKING CHANGE: Change database schema for `Section` table (and, because of the relation, the `sectionId` on `Partition`).