fix(database)!: set some columns to unique
Mark certain columns as unique in the schema if they (1) must be unique for the application to function (e.g. `Patron.loginName`), or (2) would be confusing to a user if not unique (e.g. `Section.friendlyName`).

BREAKING CHANGE: mark `Patron.loginName`, `Librarian.fullName`, `Code.content`, and `Section.friendlyName` as unique columns in the Prisma schema.