refactor(database): combine patron and librarian
Combine the Patron and Librarian models into a combined User model. This simplifies much application design. Now we don't need separate routes, separate login pages, and so on.

The combined User model doesn't have a passwordHash column, like Patron: we'll worry about passwords when authentication is implemented.

This is not a breaking change because we don't use the database in the app yet.